+++ src/addon.cc
@@ -2,8 +2,8 @@
 #include <string>
 #include <napi.h>
 #include <raylib.h>
-#include "lib/AddFunction.h"
-#include "node-raylib.h"
+#include "./lib/AddFunction.h"
+#include "./generated/node-raylib.h"
 #include "node-raymath.h"
 #include "node-rlgl.h"
 #include "wrappedfunctions.h"
